% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpQval_example.R
\name{plot.pQval}
\alias{plot.pQval}
\title{Mesa plot for polytomous attribute Q-matrix validation}
\usage{
\method{plot}{pQval}(
  x,
  item,
  type = "best",
  no.qvector = 10,
  data.label = TRUE,
  eps = "auto",
  original.q.label = FALSE,
  auto.ylim = TRUE,
  ...
)
}
\arguments{
\item{x}{model object of class \code{pQval}}

\item{item}{a vector specifying which item(s) the plots are drawn for}

\item{type}{types of the plot. It can be \code{"best"} or \code{"all"}. If \code{"best"},
for all q-vectors requiring the same number of attributes, only the one with the largest PVAF
is plotted, which means \eqn{K_j} q-vectors are plotted; If \code{"all"}, all q-vectors
will be plotted.}

\item{no.qvector}{the number of q vectors that need to be plotted when \code{type="all"}. The default is 10,
which means the 10 q vectors with the largest PVAFs are plotted.}

\item{data.label}{logical; To show data label or not?}

\item{eps}{the cutoff for PVAF. If not \code{NULL}, it must be a value between 0 and 1. A horizontal line will be drawn accordingly.}

\item{original.q.label}{logical; print the label showing the original q-vector or not?}

\item{auto.ylim}{logical; create y range automatically or not?}

\item{...}{additional arguments passed to \code{plot} function}
}
\description{
The mesa plot was first proposed by de la Torre and Ma (2016) for graphically illustrating the best q-vector(s) for each item.
The q-vector on the edge of the mesa is likely to be the best q-vector.
In this function, the mesa plot for polytomous attribute Q-matrix is developed.
}
\examples{
\dontrun{
dat <- sim18GDINA$simdat
Q <- sim18GDINA$simQ
Q[18,] <- c(0,1,1)
mod1 <- GDINA(dat = dat,Q = Q,model = "GDINA",mono.constraint = TRUE)
out<- pQval(mod,eps=0.95)
item <- c(1,18)
plot(out,item=item,type="best",eps=0.95)#'
}

}
\references{
de la Torre, J., Qiu, X.-L., & Santos, C. K. (2021). An empirical Q-matrix validation method for the polytomous G-DINA model, \emph{Psychometrika, 87}, 693–724.

de la Torre, J., & Ma, W. (2016, August). Cognitive diagnosis modeling: A general framework approach and its implementation in R. A Short Course at the Fourth Conference on Statistical Methods in Psychometrics, Columbia University, New York.
}
\seealso{
\code{\link{pQval}}
}
